requirements.txt
ecdsa==0.13
future==0.16.0
oauthlib==2.0.1
pycrypto==2.6.1
PyGithub==1.32
python-jose==1.3.2
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
tweepy==3.5.0
